requirements.txt
# Core dependencies
Flask==3.0.3
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.5.2
nltk==3.9.1
textblob==0.18.0
matplotlib==3.9.2
joblib==1.4.2
seaborn==0.13.2

# Dashboard dependencies
streamlit==1.37.0
plotly==5.22.0

# Optional AI integration
google-generativeai>=0.8.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0